cricket_detection_app/requirements.txt
streamlit
pandas
numpy
matplotlib
pillow
opencv-python-headless
scikit-learn
scikit-image